refactor: split qna api constants and types by responsibility (#70)
