+++ src/constants/qna.ts
@@ -1,0 +1,4 @@
+export const QNA_API = {
+  categories: '/qna/categories',
+  questions: '/qna/questions',
+} as const
