+++ src/types/index.ts
@@ -1,2 +1,3 @@
 // types
 export * from './api-response/category'
+export * from './api-response/qna'
